Complete cinema feature

+++ src/main/java/com/foolish/showtimeservice/controller/ShowtimeController.java
@@ -73,8 +73,7 @@ public class ShowtimeController {
     */
 
     // Kiểm tra có phải là ADMIN không.
-    IdentifyResponse iResponse = identifyService.doIdentify(request);
-    if (iResponse == null || !iResponse.getActive() || !iResponse.getRoles().equals(ROLE.ADMIN))
+    if (!identifyService.isAdmin(request))
       return ResponseEntity.ok(new ResponseError(HttpStatus.FORBIDDEN.value(), "Forbidden!"));
 
 
@@ -110,8 +109,7 @@ public class ShowtimeController {
     */
 
     // Kiểm tra có phải là ADMIN không.
-    IdentifyResponse iResponse = identifyService.doIdentify(request);
-    if (iResponse == null || !iResponse.getActive() || !iResponse.getRoles().equals(ROLE.ADMIN))
+    if (!identifyService.isAdmin(request))
       return ResponseEntity.ok(new ResponseError(HttpStatus.FORBIDDEN.value(), "Forbidden!"));
 
 
@@ -139,7 +137,12 @@ public class ShowtimeController {
   }
 
   @DeleteMapping(value = "/{id}")
-  public ResponseEntity<ResponseData> deleteShowtime(@PathVariable Integer id) {
+  public ResponseEntity<ResponseData> deleteShowtime(@PathVariable Integer id, HttpServletRequest request) {
+
+    // Kiểm tra có phải là ADMIN không.
+    if (!identifyService.isAdmin(request))
+      return ResponseEntity.ok(new ResponseError(HttpStatus.FORBIDDEN.value(), "Forbidden!"));
+
     Showtime showtime = showtimeService.getShowtimeByIdOrElseThrow(id);
     showtimeService.delete(showtime);
     return ResponseEntity.ok(new ResponseData(HttpStatus.NO_CONTENT.value(), "Success", null));
@@ -177,89 +180,6 @@ public class ShowtimeController {
   }
 
 
-  // Phương thức tạo mới Cinema.
-  @PostMapping(value = "/cinemas")
-  public ResponseEntity<ResponseData> createCinema(@RequestBody @NotNull CinemaDTO dto) {
-    /*Thống nhất là ADMIN chỉ tạo các thông tin Cinema phía dưới và không có đính kèm thêm ảnh. Vì đơn giản là tại thời điểm tạo Cinema thì không có id để cho các Banner có thể refer tới.*/
-
-    /*
-     * Cinema:
-     * {
-     *   name: String,
-     *   address: String,
-     *   province: Integer
-     * }
-     * */
-
-    Province province = null;
-    try {
-      province = provinceService.findByIdOrElseThrow(dto.getProvince().getId());
-    } catch (NullPointerException e) {
-      throw new RuntimeException("Can't find province");
-    }
-    Cinema cinema = new Cinema();
-    cinema.setName(dto.getName());
-    cinema.setAddress(dto.getAddress());
-    cinema.setProvince(province);
-    cinema.setBanners(new ArrayList<>());
-    Cinema saved = cinemaService.save(cinema);
-    if (saved == null || saved.getId() <= 0) {
-      return ResponseEntity.ok(new ResponseError(HttpStatus.BAD_REQUEST.value(), "Can't save cinema"));
-    }
-    dto = cinemaMapper.toDTO(saved);
-    return ResponseEntity.ok(new ResponseData(HttpStatus.OK.value(), "Success", dto));
-  }
-
-
-  // Hàm tìm kiếm theo tiêu chí cho Role ADMIN. Chỉ thiết kế chức năng tìm kiếm và trả về theo dạng danh sách cho Role này.
-  @PostMapping(value = "/cinemas/search")
-  public ResponseEntity<ResponseData> searchCinemasByCriteria(@RequestBody @NotNull Map<String, String> criteria, @RequestParam(name = "sort", required = false) String sort, @RequestParam(name = "pageSize", required = false) Integer pageSize, @RequestParam(name = "pageNumber", required = false) Integer pageNumber) {
-
-    int pageNum = (pageNumber != null ? pageNumber : 1) - 1;
-    int size = pageSize != null ? pageSize : 10;
-    Pageable pageable = null;
-
-    if (sort != null) {
-      // sort=id:desc,name:asc
-      List<String> list = Arrays.stream(sort.split(",")).toList();
-      List<Sort.Order> orders = new ArrayList<>();
-      for (String element : list) {
-        // Nếu fromString bị lỗi nó sẽ throw ra IllegalException và GlobalExceptionHandling sẽ catch nó trong RuntimeException.
-        orders.add(new Sort.Order(Sort.Direction.fromString(element.split(":")[1].toUpperCase()), element.split(":")[0]));
-      }
-
-      pageable = PageRequest.of(pageNum, size, Sort.by(orders));
-    } else pageable = PageRequest.of(pageNum, size);
-
-    Page<CinemaDTO> page = cinemaService.findCinemasByCriteria(criteria, pageable);
-    return ResponseEntity.ok(new ResponseData(HttpStatus.OK.value(), "Success", page));
-  }
-
-  // Phương thức lấy thông tin của một Cinema. Lưu ý, nếu như Role ADMIN thì sẽ trả về CinemaDTO có chứa ProvinceDTO.
-  @GetMapping(value = "/cinemas/{id}")
-  public ResponseEntity<ResponseData> getCinema(@PathVariable Integer id) {
-    Cinema cinema = cinemaService.getCinemaByIdOrElseThrow(id);
-    CinemaDTO data = cinemaMapper.toDTO(cinema);
-    return ResponseEntity.ok(new ResponseData(HttpStatus.OK.value(), "Success", data));
-  }
-
-  // Phương thức update Cinema
-  @Transactional
-  @PatchMapping(value = "/cinemas/{id}")
-  public ResponseEntity<ResponseData> updateCinema(@PathVariable Integer id, @RequestBody @NotNull CinemaDTO dto) {
-    Cinema cinema = cinemaService.getCinemaByIdOrElseThrow(id);
-    if (StringUtils.hasText(dto.getName())) {
-      cinema.setName(dto.getName());
-    }
-    if (StringUtils.hasText(dto.getAddress())) cinema.setAddress(dto.getAddress());
-    if (dto.getProvince() != null) cinema.setProvince(provinceService.findByIdOrElseThrow(dto.getProvince().getId()));
-
-    Cinema saved = cinemaService.save(cinema);
-    if (saved == null || saved.getId() <= 0) {
-      return ResponseEntity.ok(new ResponseError(HttpStatus.BAD_REQUEST.value(), "Can't save cinema"));
-    }
-    return ResponseEntity.ok(new ResponseData(HttpStatus.NO_CONTENT.value(), "Success", null));
-  }
 
   @Transactional
   @PostMapping(value = "/rooms")
